Add skip_join and propagate options to chord task.
This option allows the chord_unlock task to not perform the task join
when terminating.

Also don't propagate max_retries errors when propagate is set to False.

These options allows to implement an auto termination of a workflow when
tasks or their result is lost for any reason. A chord not propagating
the errors and with a predefined max_retries will eventually stop retrying
and will run the next task without joining "lost" results.